Move some buffer constants into buffer/Constants
