+++ src/common/buffer/BufferLine.test.ts
@@ -3,7 +3,8 @@
  * @license MIT
  */
 import * as chai from 'chai';
-import { BufferLine, CellData, Content, NULL_CELL_CHAR, NULL_CELL_WIDTH, NULL_CELL_CODE, DEFAULT_ATTR } from './BufferLine';
+import { NULL_CELL_CHAR, NULL_CELL_WIDTH, NULL_CELL_CODE, DEFAULT_ATTR } from 'common/buffer/Constants';
+import { BufferLine, CellData, Content } from 'common/buffer//BufferLine';
 import { CharData, IBufferLine } from '../Types';
 
 class TestBufferLine extends BufferLine {
