+++ src/common/buffer/BufferReflow.test.ts
@@ -3,7 +3,8 @@
  * @license MIT
  */
 import { assert } from 'chai';
-import { BufferLine, NULL_CELL_CHAR, NULL_CELL_WIDTH, NULL_CELL_CODE } from 'common/buffer/BufferLine';
+import { BufferLine } from 'common/buffer/BufferLine';
+import { NULL_CELL_CHAR, NULL_CELL_WIDTH, NULL_CELL_CODE } from 'common/buffer/Constants';
 import { reflowSmallerGetNewLineLengths } from 'common/buffer/BufferReflow';
 
 describe('BufferReflow', () => {
